pangloss/pyproject.toml
[project]
name = "pangloss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyhumps>=3.8.0",
    "pydantic-extra-types>=2.10.2",
    "python-ulid[pydantic]>=3.0.0",
    "fastapi[standard]>=0.115.11",
    "pydantic>=2.10.6",
    "authx>=1.4.3",
    "authx-extra>=1.2.0",
    "neo4j>=5.28.1",
    "bcrypt>=4.3.0",
    "cookiecutter>=2.6.0",
    "typer>=0.15.2",
    "rich>=13.9.4",
    "javascripthon>=0.13",
    "dill>=0.3.9",
    "jsonpatch>=1.33",
    "python-forge>=18.6.0",
    "anyio>=4.8.0",
    "structlog>=25.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "ruff>=0.9.3",
    "mypy>=1.15.0",
    "pytest-cov>=6.0.0",
    "pytest-sugar>=1.0.0",
    "pytest-asyncio>=0.25.3",
    "nest-asyncio>=1.6.0",
    "pyrefly>=0.15.1",
]

[project.scripts]
pangloss = 'pangloss.cli.main:cli'

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
filterwarnings = [
    "ignore:.*neo4j.*"
]

